feature/updated controllers to integrate validation

+++ Catalog.Host/Controllers/TypeController.cs
@@ -22,22 +22,39 @@ public class TypeController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> Add(string type, string description)
+    [ProducesResponseType(typeof(AddCarResponse<int?>), (int)HttpStatusCode.OK)]
+    public async Task<IActionResult> Add([FromBody] CreateTypeRequest request)
     {
-        var result = await _typeService.Add(type, description);
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var result = await _typeService.Add(request.TypeName, request.TypeDescription);
         return Ok(new AddCarResponse<int?>() { Id = result });
     }
 
     [HttpDelete]
     public IActionResult Delete(int itemId)
     {
+        if (itemId < 0)
+        {
+            return BadRequest("Id cannot be <0");
+        }
+
         var result = _typeService.Delete(itemId);
         return Ok();
     }
 
     [HttpPut]
-    public async Task<IActionResult> Put(TypeDto item, int itemToUpdate)
+    public async Task<IActionResult> Put([FromBody] TypeDto item, int itemToUpdate)
     {
+        var validationResult = await new TypeDtoValidator().ValidateAsync(item);
+        if (!validationResult.IsValid)
+        {
+            return BadRequest(validationResult.Errors);
+        }
+
         var result = await _typeService.Put(item, itemToUpdate);
         return Ok();
     }
